requirements.txt
Flask==1.1.2
git+git://github.com/samdobson/cookiecutter#egg=cookiecutter
PyYAML==5.3.1
python-dotenv==0.14.0
